+++ CHANGELOG.rst
@@ -205,3 +205,4 @@ Changelog
 
 * Added messages and checks for dry run support in plugins.
 * Implemented full dry run mode support in banking.
+* Implemented full dry run mode support in CSV uploader.